requirements.txt
torch~=1.11.0
matplotlib~=3.4.3
numpy~=1.20.3
scikit-learn~=0.24.2
scipy~=1.7.1
opencv-python~=4.5.5.64
pyyaml~=6.0
pillow~=8.4.0
pympi-ling~=1.70.2
tqdm~=4.62.3
intervaltree~=3.1.0
torchvision~=0.12.0
pandas~=1.3.4
yaml~=0.2.5
pytorchvideo~=0.1.5
seaborn~=0.11.2
torchsummary~=1.5.1